Refatoração do DatabaseService
- Movido `DatabaseService` e `ChangeType` do módulo `databases` para
`services`
- Ajuste dos imports do `DatabaseService`
- Ajustes de PEP8
