+++ catalogmanager/tests/test_article_services.py
@@ -1,13 +1,9 @@
-import os
 from unittest.mock import patch
 
 import pytest
 
-from catalog_persistence.databases import (
-    InMemoryDBManager,
-    DatabaseService,
-    DocumentNotFound
-)
+from catalog_persistence.databases import DocumentNotFound
+from catalog_persistence.services import DatabaseService
 from catalog_persistence.models import RecordType
 from catalogmanager.article_services import (
     ArticleServices,
@@ -16,7 +12,6 @@ from catalogmanager.article_services import (
 from catalogmanager.models.article_model import (
     Article,
 )
-from catalogmanager.models.file import File
 from catalogmanager.xml.xml_tree import (
     XMLTree
 )
